requirements.txt
streamlit-extras==0.2.7
pinecone-client==2.2.1
firebase-admin==6.1.0
langchain==0.0.184
openai==0.27.5
pandas==1.5.3
numpy==1.23.5
pip==23.1.2